+++ src/app/providers/ErrorBoundary/index.ts
@@ -1,7 +1,4 @@
 import ErrorBoundary from "./ui/ErrorBoundary"
 import { BugButton } from "./ui/BugButton"
 
-export {
-    ErrorBoundary,
-    BugButton,
-}
+export { ErrorBoundary, BugButton }
