+++ src/app/providers/StoreProvider/index.ts
@@ -1,5 +1,5 @@
 import { StoreProvider } from "./ui/StoreProvider"
 import { createReduxStore } from "./config/store"
-import { StateSchema, ThunkConfig } from "./config/StateSchema"
+import type { StateSchema, ThunkConfig } from "./config/StateSchema"
 
 export { StoreProvider, createReduxStore, StateSchema, ThunkConfig }
