+++ src/entities/Basket/index.ts
@@ -3,6 +3,6 @@ export { BasketSummary } from "./ui/BasketSummary/BasketSummary"
 
 export { BasketSummaryVariant } from "./ui/BasketSummary/BasketSummary"
 export { basketReducer } from "./model/slice/basketSlice"
-export { BasketSchema } from "./model/types/basketSchema"
+export type { BasketSchema } from "./model/types/basketSchema"
 export { basketActions } from "./model/slice/basketSlice"
 export { getBasketProducts } from "./model/selectors/basketSelectors"
