+++ src/entities/CatalogNavigation/index.ts
@@ -1,7 +1,7 @@
 export { Breadcrumbs } from "./ui/Breadcrumbs/Breadcrumbs"
 
 export { getNavigationTree } from "./model/selectors/sidebarNavigationSelectors"
-export { navigationTreeType } from "./model/types/list"
+export type { navigationTreeType } from "./model/types/list"
 export type { CatalogNavigationSchema } from "./model/types/catalogNavigationSchema"
 export { CurrentTreeItemType } from "./model/types/catalogNavigationSchema"
