+++ src/entities/ModalSlider/index.ts
@@ -1,6 +1,7 @@
 import { modalsActions, modalsReducer } from "./model/slice/modalsSlice"
 import { ModalSlider, ModalSliderDirection, ModalSliderVariant } from "./ui/ModalSlider"
-import { ModalsList, ModalsSchema } from "./model/types/modalsSchema"
+import type { ModalsSchema } from "./model/types/modalsSchema"
+import { ModalsList } from "./model/const/modalList"
 import { getModalsCurrent } from "./model/selectors/getModalsCurrent/getModalsCurrent"
 
 export {
