+++ src/entities/ModalSlider/model/const/modalList.ts
@@ -1,0 +1,8 @@
+export enum ModalsList {
+    BURGER = "burger",
+    BASKET = "basket",
+    FILTERS = "filters",
+    SORT = "sort",
+    SUCCESS = "success",
+    NONE = "",
+}
