+++ src/pages/ProductDetailsPage/index.ts
@@ -1,4 +1,4 @@
 export { ProductDetailsPage } from "./ui/ProductDetailsPage"
 
-export { ProductDetailsSchema } from "./model/types/productDetailsSchema"
+export type { ProductDetailsSchema } from "./model/types/productDetailsSchema"
 export { productDetailsReducer } from "./model/slice/productDetailsSlice"
