+++ src/pages/SubCategoryPage/index.ts
@@ -1,3 +1,3 @@
 export { SubCategoryPage } from "./ui/SubCategoryPage"
 export { subcategoryPageReducer } from "./model/slice/subcategoryPageSlice"
-export { SubCategoryPageSchema } from "./model/types/subcategoryPageSchema"
+export type { SubCategoryPageSchema } from "./model/types/subcategoryPageSchema"
