+++ src/widgets/ProductCarousel/index.ts
@@ -1,3 +1,3 @@
 export { ProductCarousel, ProductCarouselVariant } from "./ui/ProductCarousel"
 export { productCarouselReducer } from "./model/slice/productCarouselSlice"
-export { ProductCarouselSchema } from "./model/types/productDetailsSchema"
+export type { ProductCarouselSchema } from "./model/types/productDetailsSchema"
